I112 implement react router to manage webpage state (#113)
* added temporary webpage state management solution

* replace main class component with functional component

* added react-router-dom dependency, replaced state management with react router
