+++ src/FE/FEsrc/components/pages/LoginPage.tsx
@@ -1,0 +1,11 @@
+import React, { useState } from 'react';
+import styled from '@emotion/styled';
+
+const LoginPage = () => {
+    
+    return (
+        <div>LOGIN PAGE</div>
+    );
+}
+
+export default LoginPage;
