+++ src/FE/FEsrc/index.tsx
@@ -6,4 +6,4 @@ export const API_URL = 'http://localhost:8000';
 
 const container = document.getElementById('root');
 const root = createRoot(container!);
-root.render(<App message="REACH by reach" />);
+root.render(<App />);
