+++ src/FE/package.json
@@ -20,6 +20,7 @@
     "@mui/material": "^5.14.17",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-router-dom": "^6.21.3",
     "typescript": "^5.2.2"
   },
   "devDependencies": {
@@ -29,6 +30,7 @@
     "@babel/preset-typescript": "^7.23.2",
     "@types/react": "^18.2.31",
     "@types/react-dom": "^18.2.14",
+    "@types/react-router-dom": "^5.3.3",
     "babel-loader": "^9.1.3",
     "html-webpack-plugin": "^5.5.3",
     "ts-loader": "^9.5.0",
