+++ src/FE/webpack.prod.js
@@ -6,6 +6,7 @@ module.exports = merge(webpackConfig, {
     mode: "production",
     output: {
         path: path.join(__dirname, '/dist'),
-        filename: 'bundle.js'
+        filename: 'bundle.js',
+        publicPath: '/'
     }
 })
